+++ app/src/main/java/budget/application/service/domain/CategoryTypeService.java
@@ -1,5 +1,6 @@
 package budget.application.service.domain;
 
+import budget.application.common.Exceptions;
 import budget.application.db.repository.CategoryTypeRepository;
 import budget.application.model.dto.request.CategoryTypeRequest;
 import budget.application.model.dto.response.CategoryTypeResponse;
@@ -43,6 +44,12 @@ public class CategoryTypeService {
         bs -> {
           CategoryTypeRepository repo = new CategoryTypeRepository(requestId, bs);
           List<CategoryType> ctList = repo.read(ids);
+
+          if (ids.size() == 1 && ctList.isEmpty()) {
+            throw new Exceptions.NotFoundException(
+                requestId, "CategoryType", ids.getFirst().toString());
+          }
+
           return new CategoryTypeResponse(ctList, ResponseMetadata.emptyResponseMetadata());
         });
   }
@@ -54,6 +61,13 @@ public class CategoryTypeService {
         bs -> {
           CategoryTypeRepository repo = new CategoryTypeRepository(requestId, bs);
           validate(requestId, ctr);
+
+          List<CategoryType> ctList = repo.read(List.of(id));
+
+          if (ctList.isEmpty()) {
+            throw new Exceptions.NotFoundException(requestId, "CategoryType", id.toString());
+          }
+
           CategoryType ctIn = CategoryType.builder().id(id).name(ctr.name()).build();
           CategoryType ctOut = repo.update(ctIn);
           return new CategoryTypeResponse(
@@ -66,6 +80,14 @@ public class CategoryTypeService {
     return tx.execute(
         bs -> {
           CategoryTypeRepository repo = new CategoryTypeRepository(requestId, bs);
+
+          List<CategoryType> ctList = repo.read(ids);
+
+          if (ids.size() == 1 && ctList.isEmpty()) {
+            throw new Exceptions.NotFoundException(
+                requestId, "CategoryType", ids.getFirst().toString());
+          }
+
           int deleteCount = repo.delete(ids);
           return new CategoryTypeResponse(
               List.of(), ResponseMetadataUtils.defaultDeleteResponseMetadata(deleteCount));
@@ -74,11 +96,11 @@ public class CategoryTypeService {
 
   private void validate(String requestId, CategoryTypeRequest ctr) {
     if (ctr == null) {
-      throw new IllegalArgumentException(
+      throw new Exceptions.BadRequestException(
           String.format("[%s] Category type request cannot be null...", requestId));
     }
     if (CommonUtilities.isEmpty(ctr.name())) {
-      throw new IllegalArgumentException(
+      throw new Exceptions.BadRequestException(
           String.format("[%s] Category type name cannot be empty...", requestId));
     }
   }
